refactor: change tooling/i18n by infrastructure/i18n
