+++ api/tests/certification/enrolment/integration/infrastructure/ods/read-ods-utils_test.js
@@ -6,7 +6,6 @@ const { readFile } = promises;
 
 import _ from 'lodash';
 
-import { getI18n } from '../../../../../tooling/i18n/i18n.js';
 const i18n = getI18n();
 
 import * as url from 'node:url';
@@ -19,6 +18,7 @@ import {
   validateOdsHeaders,
 } from '../../../../../../src/certification/enrolment/infrastructure/utils/ods/read-ods-utils.js';
 import { UnprocessableEntityError } from '../../../../../../src/shared/application/http-errors.js';
+import { getI18n } from '../../../../../../src/shared/infrastructure/i18n/i18n.js';
 import { catchErr, expect } from '../../../../../test-helper.js';
 const __dirname = url.fileURLToPath(new URL('.', import.meta.url));
