+++ api/tests/certification/enrolment/unit/application/attendance-sheet-controller_test.js
@@ -1,7 +1,7 @@
 import { attendanceSheetController } from '../../../../../src/certification/enrolment/application/attendance-sheet-controller.js';
 import { usecases } from '../../../../../src/certification/enrolment/domain/usecases/index.js';
+import { getI18n } from '../../../../../src/shared/infrastructure/i18n/i18n.js';
 import { expect, hFake, sinon } from '../../../../test-helper.js';
-import { getI18n } from '../../../../tooling/i18n/i18n.js';
 
 describe('Unit | Controller | attendance-sheet-controller', function () {
   describe('#getAttendanceSheet', function () {
