+++ api/tests/certification/enrolment/unit/domain/usecases/validate-sessions_test.js
@@ -7,8 +7,8 @@ import { CERTIFICATION_SESSIONS_ERRORS } from '../../../../../../src/certificati
 import { CpfBirthInformationValidation } from '../../../../../../src/certification/shared/domain/services/certification-cpf-service.js';
 import { CERTIFICATION_CENTER_TYPES } from '../../../../../../src/shared/domain/constants.js';
 import { CertificationCandidate } from '../../../../../../src/shared/domain/models/index.js';
+import { getI18n } from '../../../../../../src/shared/infrastructure/i18n/i18n.js';
 import { domainBuilder, expect, sinon } from '../../../../../test-helper.js';
-import { getI18n } from '../../../../../tooling/i18n/i18n.js';
 
 const userId = 1234;
 const cachedValidatedSessionsKey = 'uuid';
