+++ api/tests/certification/results/integration/infrastructure/utils/csv/certification-results/get-session-certification-results-csv_test.js
@@ -1,7 +1,8 @@
 import { getSessionCertificationResultsCsv } from '../../../../../../../../src/certification/results/infrastructure/utils/csv/certification-results/get-session-certification-results-csv.js';
 import { AutoJuryCommentKeys } from '../../../../../../../../src/certification/shared/domain/models/JuryComment.js';
+import { getI18n } from '../../../../../../../../src/shared/infrastructure/i18n/i18n.js';
 import { domainBuilder, expect } from '../../../../../../../test-helper.js';
-import { getI18n } from '../../../../../../../tooling/i18n/i18n.js';
+
 const i18n = getI18n();
 const translate = i18n.__;
