+++ api/tests/certification/results/unit/application/certification-controller_test.js
@@ -1,8 +1,8 @@
 import { certificationController } from '../../../../../src/certification/results/application/certification-controller.js';
 import { usecases } from '../../../../../src/certification/results/domain/usecases/index.js';
 import { SESSIONS_VERSIONS } from '../../../../../src/certification/shared/domain/models/SessionVersion.js';
+import { getI18n } from '../../../../../src/shared/infrastructure/i18n/i18n.js';
 import { domainBuilder, expect, hFake, sinon } from '../../../../test-helper.js';
-import { getI18n } from '../../../../tooling/i18n/i18n.js';
 
 describe('Certification | Results | Unit | Application | certifications-controller', function () {
   describe('#getCertificationByVerificationCode', function () {
