+++ api/tests/certification/results/unit/application/certification-results-controller_test.js
@@ -1,7 +1,7 @@
 import { certificationResultsController } from '../../../../../src/certification/results/application/certification-results-controller.js';
 import { usecases } from '../../../../../src/certification/results/domain/usecases/index.js';
+import { getI18n } from '../../../../../src/shared/infrastructure/i18n/i18n.js';
 import { domainBuilder, expect, hFake, sinon } from '../../../../test-helper.js';
-import { getI18n } from '../../../../tooling/i18n/i18n.js';
 
 describe('Certification | Results | Unit | Controller | certification results', function () {
   describe('#getCleaCertifiedCandidateDataCsv', function () {
