+++ api/tests/certification/results/unit/application/organization-controller_test.js
@@ -1,7 +1,7 @@
 import { organizationController } from '../../../../../src/certification/results/application/organization-controller.js';
 import { usecases } from '../../../../../src/certification/results/domain/usecases/index.js';
+import { getI18n } from '../../../../../src/shared/infrastructure/i18n/i18n.js';
 import { domainBuilder, expect, hFake, sinon } from '../../../../test-helper.js';
-import { getI18n } from '../../../../tooling/i18n/i18n.js';
 
 describe('Certification | Course | Unit | Application | Organizations | organization-controller', function () {
   describe('#downloadCertificationResults', function () {
