+++ api/tests/certification/results/unit/infrastructure/serializers/private-certificate-serializer_test.js
@@ -3,8 +3,8 @@ import { ResultCompetenceTree } from '../../../../../../src/certification/result
 import * as serializer from '../../../../../../src/certification/results/infrastructure/serializers/private-certificate-serializer.js';
 import { AutoJuryCommentKeys } from '../../../../../../src/certification/shared/domain/models/JuryComment.js';
 import { SESSIONS_VERSIONS } from '../../../../../../src/certification/shared/domain/models/SessionVersion.js';
+import { getI18n } from '../../../../../../src/shared/infrastructure/i18n/i18n.js';
 import { domainBuilder, expect } from '../../../../../test-helper.js';
-import { getI18n } from '../../../../../tooling/i18n/i18n.js';
 
 describe('Certification | Results | Unit | Infrastructure | Serializers | private-certificate-serializer', function () {
   let translate;
