+++ api/tests/certification/results/unit/infrastructure/utils/csv/certification-results/CertificationResultsCsvValues_test.js
@@ -1,8 +1,8 @@
 import { CertificationResultsCsvValues } from '../../../../../../../../src/certification/results/infrastructure/utils/csv/certification-results/CertificationResultsCsvValues.js';
 import { AutoJuryCommentKeys } from '../../../../../../../../src/certification/shared/domain/models/JuryComment.js';
 import { CertificationResult } from '../../../../../../../../src/shared/domain/models/index.js';
+import { getI18n } from '../../../../../../../../src/shared/infrastructure/i18n/i18n.js';
 import { domainBuilder, expect } from '../../../../../../../test-helper.js';
-import { getI18n } from '../../../../../../../tooling/i18n/i18n.js';
 
 describe('Unit | Infrastructure | Utils | Csv | CertificationResultsCsvValues', function () {
   let i18n, translate;
