+++ api/tests/certification/session-management/integration/infrastructure/utils/pdf/attendance-sheet-pdf_test.js
@@ -6,9 +6,9 @@ import i18n from 'i18n';
 import pdfLibUtils from 'pdf-lib/cjs/utils/index.js';
 
 import { getAttendanceSheetPdfBuffer } from '../../../../../../../src/certification/enrolment/infrastructure/utils/pdf/attendance-sheet-pdf.js';
+import { getI18n } from '../../../../../../../src/shared/infrastructure/i18n/i18n.js';
 import { domainBuilder, expect, sinon } from '../../../../../../test-helper.js';
 import { isSameBinary } from '../../../../../../tooling/binary-comparator.js';
-import { getI18n } from '../../../../../../tooling/i18n/i18n.js';
 const __dirname = url.fileURLToPath(new URL('.', import.meta.url));
 
 const directory = path.resolve(path.join(__dirname, '../../../../../../../translations'));
