+++ api/tests/certification/session-management/unit/application/invigilator-kit-controller_test.js
@@ -1,8 +1,8 @@
 import { invigilatorKitController } from '../../../../../src/certification/session-management/application/invigilator-kit-controller.js';
 import { usecases } from '../../../../../src/certification/session-management/domain/usecases/index.js';
 import { LANGUAGES_CODE } from '../../../../../src/shared/domain/services/language-service.js';
+import { getI18n } from '../../../../../src/shared/infrastructure/i18n/i18n.js';
 import { domainBuilder, expect, hFake, sinon } from '../../../../test-helper.js';
-import { getI18n } from '../../../../tooling/i18n/i18n.js';
 
 describe('Certification | Session Management | Unit | Application | Controller | Invigilator Kit', function () {
   describe('#getInvigilatorKitPdf', function () {
@@ -19,13 +19,12 @@ describe('Certification | Session Management | Unit | Application | Controller |
         const sessionMainInfo = domainBuilder.certification.sessionManagement.buildSession({ id: 1 });
         const invigilatorKitBuffer = 'binary string';
         const userId = 1;
-        const i18n = getI18n();
+        const i18n = getI18n(lang);
         const request = {
           i18n,
           auth: { credentials: { userId } },
           params: { sessionId: sessionMainInfo.id },
         };
-        i18n.setLocale(lang);
 
         const invigilatorKitPdf = {
           getInvigilatorKitPdfBuffer: sinon.stub(),
