+++ api/tests/certification/session-management/unit/application/jury-certification-controller_test.js
@@ -1,7 +1,7 @@
 import { juryCertificationController } from '../../../../../src/certification/session-management/application/jury-certification-controller.js';
 import { usecases } from '../../../../../src/certification/session-management/domain/usecases/index.js';
+import { getI18n } from '../../../../../src/shared/infrastructure/i18n/i18n.js';
 import { domainBuilder, expect, hFake, sinon } from '../../../../test-helper.js';
-import { getI18n } from '../../../../tooling/i18n/i18n.js';
 
 describe('Certification | Session-management | Unit | Application | jury-certification-controller', function () {
   describe('#getJuryCertification', function () {
