+++ api/tests/certification/shared/unit/domain/models/JuryComment_test.js
@@ -3,8 +3,8 @@ import {
   JuryComment,
   JuryCommentContexts,
 } from '../../../../../../src/certification/shared/domain/models/JuryComment.js';
+import { getI18n } from '../../../../../../src/shared/infrastructure/i18n/i18n.js';
 import { expect } from '../../../../../test-helper.js';
-import { getI18n } from '../../../../../tooling/i18n/i18n.js';
 
 describe('Unit | Domain | Models | JuryComment', function () {
   let translate;
