+++ api/tests/integration/domain/usecases/create-and-reconcile-user-to-organization-learner_test.js
@@ -21,8 +21,8 @@ import { cryptoService } from '../../../../src/shared/domain/services/crypto-ser
 import * as userService from '../../../../src/shared/domain/services/user-service.js';
 import * as passwordValidator from '../../../../src/shared/domain/validators/password-validator.js';
 import * as userValidator from '../../../../src/shared/domain/validators/user-validator.js';
+import { getI18n } from '../../../../src/shared/infrastructure/i18n/i18n.js';
 import { catchErr, databaseBuilder, expect } from '../../../test-helper.js';
-import { getI18n } from '../../../tooling/i18n/i18n.js';
 
 const i18n = getI18n();
