+++ api/tests/integration/domain/usecases/create-organizations-with-tags-and-target-profiles_test.js
@@ -119,7 +119,7 @@ describe('Integration | UseCases | create-organizations-with-tags-and-target-pro
           },
           {
             attribute: 'locale',
-            message: "La locale doit avoir l'une des valeurs suivantes : en, es, fr, fr-be, fr-fr, nl-be",
+            message: "La locale doit avoir l'une des valeurs suivantes : en, es, fr, fr-be, fr-fr, nl-be, nl",
           },
           {
             attribute: 'locale',
