+++ api/tests/integration/infrastructure/plugins/i18n_test.js
@@ -1,6 +1,6 @@
+import { getI18n } from '../../../../src/shared/infrastructure/i18n/i18n.js';
 import * as i18nPlugin from '../../../../src/shared/infrastructure/plugins/i18n.js';
 import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
-import { getI18n } from '../../../tooling/i18n/i18n.js';
 
 describe('Integration | Infrastructure | plugins | i18n', function () {
   let httpTestServer;
