+++ api/tests/prescription/learner-management/acceptance/application/sup-organization-management-route_test.js
@@ -1,12 +1,12 @@
 import { SupOrganizationLearnerImportHeader } from '../../../../../src/prescription/learner-management/infrastructure/serializers/csv/sup-organization-learner-import-header.js';
 import { Membership } from '../../../../../src/shared/domain/models/Membership.js';
+import { getI18n } from '../../../../../src/shared/infrastructure/i18n/i18n.js';
 import {
   createServer,
   databaseBuilder,
   expect,
   generateValidRequestAuthorizationHeader,
 } from '../../../../test-helper.js';
-import { getI18n } from '../../../../tooling/i18n/i18n.js';
 
 const i18n = getI18n();
 const supOrganizationLearnerImportHeader = new SupOrganizationLearnerImportHeader(i18n).columns
