+++ api/tests/prescription/learner-management/unit/domain/models/SupOrganizationLearnerSet_test.js
@@ -1,7 +1,7 @@
 import { SupOrganizationLearnerSet } from '../../../../../../src/prescription/learner-management/domain/models/SupOrganizationLearnerSet.js';
 import { DomainError, EntityValidationError } from '../../../../../../src/shared/domain/errors.js';
+import { getI18n } from '../../../../../../src/shared/infrastructure/i18n/i18n.js';
 import { catchErr, expect } from '../../../../../test-helper.js';
-import { getI18n } from '../../../../../tooling/i18n/i18n.js';
 
 describe('Unit | Models | SupOrganizationLearnerSet', function () {
   let i18n, learnerSet;
