+++ api/tests/prescription/learner-management/unit/domain/usecases/get-organization-learners-csv-template_test.js
@@ -3,8 +3,8 @@ import _ from 'lodash';
 import { OrganizationLearnerImportFormat } from '../../../../../../src/prescription/learner-management/domain/models/OrganizationLearnerImportFormat.js';
 import { getOrganizationLearnersCsvTemplate } from '../../../../../../src/prescription/learner-management/domain/usecases/get-organization-learners-csv-template.js';
 import { UserNotAuthorizedToAccessEntityError } from '../../../../../../src/shared/domain/errors.js';
+import { getI18n } from '../../../../../../src/shared/infrastructure/i18n/i18n.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../../../test-helper.js';
-import { getI18n } from '../../../../../tooling/i18n/i18n.js';
 
 describe('Unit | UseCase | get-organization-learners-csv-template', function () {
   let userId, organizationId, membershipRepository, organizationLearnerImportFormatRepository, i18n;
