+++ api/tests/prescription/learner-management/unit/domain/usecases/import-sup-organization-learners_test.js
@@ -1,8 +1,9 @@
 import { OrganizationImport } from '../../../../../../src/prescription/learner-management/domain/models/OrganizationImport.js';
 import { importSupOrganizationLearners } from '../../../../../../src/prescription/learner-management/domain/usecases/import-sup-organization-learners.js';
 import { SupOrganizationLearnerImportHeader } from '../../../../../../src/prescription/learner-management/infrastructure/serializers/csv/sup-organization-learner-import-header.js';
+import { getI18n } from '../../../../../../src/shared/infrastructure/i18n/i18n.js';
 import { catchErr, expect, sinon, toStream } from '../../../../../test-helper.js';
-import { getI18n } from '../../../../../tooling/i18n/i18n.js';
+
 const i18n = getI18n();
 
 const supOrganizationLearnerImportHeader = new SupOrganizationLearnerImportHeader(i18n).columns
