+++ api/tests/unit/domain/models/CertificationCandidate_test.js
@@ -1,6 +1,6 @@
 import { CertificationCandidate } from '../../../../src/shared/domain/models/index.js';
+import { getI18n } from '../../../../src/shared/infrastructure/i18n/i18n.js';
 import { domainBuilder, expect } from '../../../test-helper.js';
-import { getI18n } from '../../../tooling/i18n/i18n.js';
 
 const i18n = getI18n();
 const translate = i18n.__;
