+++ api/tests/unit/domain/services/mail-service_test.js
@@ -2,6 +2,7 @@ import * as mailService from '../../../../lib/domain/services/mail-service.js';
 import { config as settings } from '../../../../src/shared/config.js';
 import { LOCALE } from '../../../../src/shared/domain/constants.js';
 import { tokenService } from '../../../../src/shared/domain/services/token-service.js';
+import { getI18n } from '../../../../src/shared/infrastructure/i18n/i18n.js';
 import { urlBuilder } from '../../../../src/shared/infrastructure/utils/url-builder.js';
 import { mailer } from '../../../../src/shared/mail/infrastructure/services/mailer.js';
 import en from '../../../../translations/en.json' with { type: 'json' };
@@ -9,7 +10,6 @@ import fr from '../../../../translations/fr.json' with { type: 'json' };
 import { es } from '../../../../translations/index.js';
 import nl from '../../../../translations/nl.json' with { type: 'json' };
 import { expect, sinon } from '../../../test-helper.js';
-import { getI18n } from '../../../tooling/i18n/i18n.js';
 
 const mainTranslationsMapping = {
   fr,
