+++ api/tests/unit/domain/services/session-publication-service_test.js
@@ -8,8 +8,8 @@ import {
 } from '../../../../src/shared/domain/errors.js';
 import { status } from '../../../../src/shared/domain/models/AssessmentResult.js';
 import { EmailingAttempt } from '../../../../src/shared/domain/models/index.js';
+import { getI18n } from '../../../../src/shared/infrastructure/i18n/i18n.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
-import { getI18n } from '../../../tooling/i18n/i18n.js';
 
 describe('Unit | UseCase | session-publication-service', function () {
   const sessionId = 123;
