+++ api/tests/unit/domain/validators/organization-with-tags-and-target-profiles-script_test.js
@@ -100,7 +100,7 @@ describe('Unit | Domain | Validators | organization-with-tags-and-target-profile
         expect(error.message).to.equal(`Échec de validation de l'entité.`);
         expect(error.invalidAttributes).to.deep.include({
           attribute: 'locale',
-          message: `La locale doit avoir l'une des valeurs suivantes : en, es, fr, fr-be, fr-fr, nl-be`,
+          message: `La locale doit avoir l'une des valeurs suivantes : en, es, fr, fr-be, fr-fr, nl-be, nl`,
         });
       });
     });
